instant upvote update
